# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Library Management System API",
    "description": "API for the Library Management System",
    "contact": {
      "name": "Petr Boháč",
      "email": "petrbohac3@seznam.cz"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "Local host API server"
    }
  ],
...
  "paths": {
    "/statistics": {
      "get": {
        "description": "System statistics for dashboards",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get all users",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Get all books",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "description": "Get all books",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/authors": {
      "get": {
        "tags": [
          "Authors"
        ],
        "description": "Get all authors",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Authors"
        ],
        "description": "Create a new author",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/history": {
      "get": {
        "tags": [
          "History"
        ],
        "description": "Get all history",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "History"
        ],
        "description": "Create a new history",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid book/user/date"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}